+++ NEWS.md
@@ -1,3 +1,16 @@
+# Development version for bugs fixes or added features
+* v.1.1.0.900
+This addresses bugs and adds some minor features. Values returned from  spreval::eff, i.e.,
+"appeff"" and "appadeq" have been corrected.  Most functions that output plots, or allow optional plots
+(i.e., eda.shape, plotss, sfplot, rotecdf) now allow optional graphical parameters to be passed via the ... convention.
+A vignette to illustrate use of the package in landscape sprinkler irrigation application has been added, along with an 
+accompanying landscape irrigation data added to the catchcan dataset as "catchcan$landscape". 
+---
+
+# Initial released version in CRAN (2021-07-05)
+* v.1.0.0
+---
+
 # spreval 1.0.0
 * version to submit to CRAN
 [package in github](https://github.com/glgrabow/spreval/blob/master/packages/spreval_1.0.0.tar.gz)